+++ app/helpers/dorsale/all_helpers.rb
@@ -3,8 +3,10 @@ module Dorsale
     include ButtonHelper
     include CommentsHelper
     include ContextHelper
+    include FiltersHelper
     include FormHelper
     include LinkHelper
+    include PaginationHelper
     include TextHelper
   end
 end