+++ app/helpers/dorsale/pagination_helper.rb
@@ -1,0 +1,8 @@
+module Dorsale
+  module PaginationHelper
+    def paginate(objects, options = {})
+      options = {theme: "twitter-bootstrap-3"}.merge(options)
+      super(objects, options).gsub(/>(\s+)</, '><').html_safe
+    end
+  end
+end