+++ lib/dorsale/engine.rb
@@ -13,6 +13,7 @@ require "bootstrap-kaminari-views"
 require "rails-i18n"
 require "cancan"
 require "awesome_print"
+require "kaminari-i18n"
 
 require "dorsale/simple_form"
 require "dorsale/simple_form_bootstrap"